Upgrade AuditLog service to .NET 6
