+++ src/AuditlogService/AuditlogManagerConfig.cs
@@ -1,7 +1,6 @@
-namespace AuditlogService
+namespace Pitstop.AuditlogService;
+
+public class AuditlogManagerConfig
 {
-    public class AuditlogManagerConfig
-    {
-        public string LogPath { get; set; }
-    }
+    public string LogPath { get; set; }
 }
