+++ src/AuditlogService/Dockerfile
@@ -1,17 +1,16 @@
-FROM mcr.microsoft.com/dotnet/sdk:5.0 AS build-env
+FROM mcr.microsoft.com/dotnet/sdk:6.0-bullseye-slim AS build-env
 WORKDIR /app
 
 # Copy necessary files and restore as distinct layer
 COPY *.csproj ./
 RUN dotnet restore
 
-
 # Copy everything else and build
 COPY . ./
 RUN dotnet publish -c Release -o out
 
 # Build runtime image
-FROM mcr.microsoft.com/dotnet/runtime:5.0
+FROM mcr.microsoft.com/dotnet/runtime:6.0-bullseye-slim
 COPY --from=build-env /app/out .
 
 # Start
