+++ src/AuditlogService/GlobalUsings.cs
@@ -1,0 +1,6 @@
+global using Pitstop.AuditlogService;
+global using Pitstop.Infrastructure.Messaging;
+global using Pitstop.Infrastructure.Messaging.Configuration;
+global using Serilog;
+global using Microsoft.Extensions.Hosting;
+global using Microsoft.Extensions.DependencyInjection;
